requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
sqlalchemy==1.4.27
psycopg2-binary==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.2
stripe==2.64.0

# Testing dependencies (TDD required)
pytest==7.2.0
pytest-asyncio==0.20.2
pytest-cov==4.0.0
httpx==0.23.1
